Add battery chargers switches to UI
For now let's keep them in the monitoring tab. Don't really have a
proper place for them. Ultimately it doesn't matter much as I just added
them to EDI for the sake of completeness. I'm mostly just using the
button on the plug itself and, frankly, I added them just because the
chargers have no switch and unplugging the cable is annoying.